requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
amqp==5.0.9
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.6.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.3
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==3.2.12
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-timezone-field==4.2.3
djangorestframework==3.13.1
frozenlist==1.3.0
gunicorn==20.1.0
hiredis==2.0.0
idna==3.3
kombu==5.2.3
multidict==6.0.2
packaging==21.3
prettytable==3.1.1
prompt-toolkit==3.0.27
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pyparsing==3.0.7
python-crontab==2.6.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
redis==4.1.3
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
telepot==12.7
toml==0.10.2
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
yarl==1.7.2
